Implement LambdaColumn and base ImageColumn on it (#99)
CellColumns create a Python object for each cell. This leads to very slow column creation for columns with many rows. LambdaColumns provide some of the benefits of a cell column (apply arbitrary loading and transformation code lazily) by wrapping an existing column or DataPanel, and simply applying the transform to each row.

ImageColumn now subclasses LambdaColumn, so is substantially faster to create.

Closes #97.

TODO:
[] We are planning a refactor of Meerkat IO #98, so I'm holding off thoroughly testing the IO capabilities of LambdaColumn