Make column and DataPanel testing more modular (#118)
Closes #109
The BlockManager (see #104) introduces a need for more robust DataPanel testing that tests DataPanels with a diverse set of columns. As we add more columns, we don't want to have to update the DataPanel tests for each new column. Instead, we should specify a TestBed for each column that plugs in to the DataPanel tests.

Started this for NumpyArrayColumn with #108.

Co-authored-by: Priya <priyamis@cse.iitk.ac.in>